+++ packages/security/reducers.js
@@ -1,15 +1,15 @@
-import { createAction, createReducer } from 'redux-act'
-import Fingerprint2 from 'fingerprintjs2'
-import Storage from '@qasir/storage-local'
-import Crypto from '@qasir/security/crypto'
-import Variables from '@qasir/security/variables'
-import Repository  from '@qasir/auth/repository'
+import { createAction, createReducer } from 'redux-act';
+import Fingerprint2 from 'fingerprintjs2';
+import Storage from '@qasir/storage-local';
+import Crypto from '@qasir/security/crypto';
+import Variables from '@qasir/security/variables';
+import Repository  from '@qasir/auth/repository';
 
-export const setClientSecret = createAction("@@QASIR_CRM_SET_CLIENT_SECRET")
+export const setClientSecret = createAction("@@QASIR_CRM_SET_CLIENT_SECRET");
 
 export const generateClientSecret = () => (dispatch) => {
-  const crypto = new Crypto
-  const storage = new Storage
+  const crypto = new Crypto;
+  const storage = new Storage;
 
   // handle for get device id on browser using
   // fingerprint javascript
@@ -19,7 +19,7 @@ export const generateClientSecret = () => (dispatch) => {
       const repository = new Repository
       repository.serviceGenerateClientSecret((flag) => {
           console.log('success client secret')
-      })
+      });
   })
 }
 
@@ -29,4 +29,4 @@ const initialState = {
 
 export default createReducer({
   [setClientSecret]: (state, clientSecret) => ({...state, clientSecret})
-}, initialState)
+}, initialState);