+++ packages/storage-local/index.js
@@ -1,19 +1,19 @@
 class storage {
 	constructor() {
-		this.mystorage = window.localStorage
+		this.mystorage = window.localStorage;
 	}
 
 	setData(key, data) {
-		this.mystorage.setItem(key, data)
+		this.mystorage.setItem(key, data);
 	}
 
 	getData(key) {
-		return this.mystorage.getItem(key)
+		return this.mystorage.getItem(key);
 	}
 
 	removeData(key) {
-		this.mystorage.removeItem(key)
+		this.mystorage.removeItem(key);
 	}
 }
 
-export default storage
+export default storage;